src/requirements.txt
discord.py
requests
discord-py-slash-command
python-dotenv
lavalink
praw
psycopg2-binary
beautifulsoup4
Pillow